chore(query-engine): extract AgentMdPrService with typed error hierarchy
